+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.0.15 - October, 22, 2019
 ---------------------
 - Fix: Stream gallery uses full post message as image description
+- Fix #20: Missing translation in gallery edit form
 
 1.0.14 - October, 16, 2019
 ---------------------